move crashreport initialization
