+++ src/MarkdownEdit/App.xaml.cs
@@ -43,10 +43,7 @@ namespace MarkdownEdit
 
         private void OnStartup(object sender, StartupEventArgs ea)
         {
-            AppDomain.CurrentDomain.UnhandledException += (o, args) => ShowCrashReport((Exception)args.ExceptionObject);
-            Current.DispatcherUnhandledException += (o, args) => ShowCrashReport(args.Exception);
-            TaskScheduler.UnobservedTaskException += (o, args) => ShowCrashReport(args.Exception);
-
+            CrashReport.Initialize(this);
             InitializeSettings();
 
             if (UserSettings == null || AlreadyEditingFile())
@@ -92,13 +89,6 @@ namespace MarkdownEdit
             }
 
             return false;
-        }
-
-        private static void ShowCrashReport(Exception exception)
-        {
-            var crashReport = new CrashReport(exception);
-            MessageBox.Show(crashReport.Report);
-            Environment.Exit(110);
         }
 
         private static void InitializeSettings()
